requirements.txt
django==5.1.5
requests==2.32.3
python-dotenv==1.0.1
